@@ -1,6 +1,6 @@
 import 'package:location/location.dart';
-import 'package:mapbox_gl/mapbox_gl.dart';
+import 'package:sport_log/helpers/lat_lng.dart';
 
 extension LocationDataExtension on LocationData {
-  LatLng get latLng => LatLng(latitude!, longitude!);
+  LatLng get latLng => LatLng(lat: latitude!, lng: longitude!);
 }
