@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mapbox_gl/mapbox_gl.dart';
+import 'package:mapbox_maps_flutter/mapbox_maps_flutter.dart';
 import 'package:sport_log/helpers/extensions/map_controller_extension.dart';
 import 'package:sport_log/widgets/app_icons.dart';
 
@@ -9,7 +9,7 @@ class SetNorthButton extends StatelessWidget {
     super.key,
   });
 
-  final MapboxMapController mapController;
+  final MapboxMap mapController;
 
   @override
   Widget build(BuildContext context) {
